requirements.txt
streamlit>=1.30.0
google-generativeai>=0.4.0
plotly>=5.18.0
streamlit-drawable-canvas>=0.9.3
Pillow>=10.0.0
